+++ CHANGELOG.md
@@ -52,7 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Improved performance of antenna metrics calculation by utilizing cached wave amplitude calculations instead of recomputing wave amplitudes for each port excitation in the `TerminalComponentModelerData`.
 - Changed hashing method in `Tidy3dBaseModel` from sha256 to md5.
-- Allowing for more geometries in a ClipOperation geometry.
+- Removed validators on limiting the number of geometries in a ClipOperation geometry.
 - Improved the speed of computing `Box` shape derivatives when used inside a `GeometryGroup`.
 - All RF and microwave specific components now inherit from `MicrowaveBaseModel`.
 - `DirectivityMonitor` now forces `far_field_approx` to `True`, which was previously configurable.